+++ lib/TagCache.php
@@ -2,7 +2,7 @@
 /**
  * Wei Framework
  *
- * @copyright   Copyright (c) 2008-2015 Twin Huang
+ * @copyright   Copyright (c) 2008-2016 Twin Huang
  * @license     http://opensource.org/licenses/mit-license.php MIT License
  */
 namespace Wei;