+++ lib/Wei.php
@@ -2,7 +2,7 @@
 /**
  * Wei Framework
  *
- * @copyright   Copyright (c) 2008-2015 Twin Huang
+ * @copyright   Copyright (c) 2008-2016 Twin Huang
  * @license     http://opensource.org/licenses/mit-license.php MIT License
  */
 
@@ -272,7 +272,7 @@ namespace Wei
         /**
          * Version
          */
-        const VERSION = '0.9.20';
+        const VERSION = '0.9.21';
 
         /**
          * The configurations for all objects